WIP: move prevote.go and request.go into core.go
