--- consensus/tendermint/core/request.go
@@ -1,100 +1,0 @@
-// Copyright 2017 The go-ethereum Authors
-// This file is part of the go-ethereum library.
-//
-// The go-ethereum library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The go-ethereum library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
-
-package core
-
-import "github.com/clearmatics/autonity/consensus/tendermint"
-
-func (c *core) handleRequest(request *tendermint.Request) error {
-	logger := c.logger.New("state", c.state, "seq", c.current.sequence)
-
-	if err := c.checkRequestMsg(request); err != nil {
-		if err == errInvalidMessage {
-			logger.Warn("invalid request")
-			return err
-		}
-		logger.Warn("unexpected request", "err", err, "number", request.ProposalBlock.Number(), "hash", request.ProposalBlock.Hash())
-		return err
-	}
-
-	logger.Trace("handleRequest", "number", request.ProposalBlock.Number(), "hash", request.ProposalBlock.Hash())
-
-	c.latestPendingRequest = request
-	// TODO: remove, we should not be sending a proposal from handleRequest
-	if c.state == StateAcceptRequest {
-		c.sendProposal(request)
-	}
-	return nil
-}
-
-// check request state
-// return errInvalidMessage if the message is invalid
-// return errFutureMessage if the sequence of proposal is larger than current sequence
-// return errOldMessage if the sequence of proposal is smaller than current sequence
-func (c *core) checkRequestMsg(request *tendermint.Request) error {
-	if request == nil || request.ProposalBlock == nil {
-		return errInvalidMessage
-	}
-
-	if c := c.current.sequence.Cmp(request.ProposalBlock.Number()); c > 0 {
-		return errOldMessage
-	} else if c < 0 {
-		return errFutureMessage
-	} else {
-		return nil
-	}
-}
-
-func (c *core) storeRequestMsg(request *tendermint.Request) {
-	logger := c.logger.New("state", c.state)
-
-	logger.Trace("Store future request", "number", request.ProposalBlock.Number(), "hash", request.ProposalBlock.Hash())
-
-	c.pendingRequestsMu.Lock()
-	defer c.pendingRequestsMu.Unlock()
-
-	c.pendingRequests.Push(request, float32(-request.ProposalBlock.Number().Int64()))
-}
-
-func (c *core) processPendingRequests() {
-	c.pendingRequestsMu.Lock()
-	defer c.pendingRequestsMu.Unlock()
-
-	for !(c.pendingRequests.Empty()) {
-		m, prio := c.pendingRequests.Pop()
-		r, ok := m.(*tendermint.Request)
-		if !ok {
-			c.logger.Warn("Malformed request, skip", "msg", m)
-			continue
-		}
-		// Push back if it's a future message
-		err := c.checkRequestMsg(r)
-		if err != nil {
-			if err == errFutureMessage {
-				c.logger.Trace("Stop processing request", "number", r.ProposalBlock.Number(), "hash", r.ProposalBlock.Hash())
-				c.pendingRequests.Push(m, prio)
-				break
-			}
-			c.logger.Trace("Skip the pending request", "number", r.ProposalBlock.Number(), "hash", r.ProposalBlock.Hash(), "err", err)
-			continue
-		}
-		c.logger.Trace("Post pending request", "number", r.ProposalBlock.Number(), "hash", r.ProposalBlock.Hash())
-
-		go c.sendEvent(tendermint.RequestEvent{
-			ProposalBlock: r.ProposalBlock,
-		})
-	}
-}
